backend/ms-auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
# Dependências de testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
asyncio==3.4.3
faker==20.0.0
pytest-mock==3.12.0
pytest-dependency==0.5.1
email-validator==2.1.0
factory-boy==3.3.0 